pipeline/rss/requirements.txt
feedparser
pytest
requests
logger
openai
python-dotenv
boto3
pyarrow
pandas
psycopg2-binary
